requirements.txt
Flask>=3.0.0,<4
# WSGI server for Render
gunicorn>=21.2.0,<22
# ML/DS deps
tensorflow-cpu>=2.15,<3
numpy>=1.26,<3
pandas>=2.2,<3
scikit-learn>=1.4,<2
joblib>=1.3,<2
Pillow>=10,<11
